[frontend] fix(use-case): rename hook files to kebab-case (#2374)
