+++ apps/frontend/graphql/use-case/hooks/use-delete-use-case.ts
